QuickLinksCommand: Make deleting the request message optional

+++ src/commands/quickLinksCommand.ts
@@ -3,35 +3,40 @@ import { CommandParser } from "../models/commandParser";
 import assert from "assert";
 
 export class QuickLinksCommand implements Command {
-    readonly links: { help: string; response: string; name: string }[] = [
+    readonly links: { help: string; response: string; name: string; deleteRequest: boolean }[] = [
         {
             name: "faq",
             response: "FAQ: http://serenityos.org/faq/",
             help: "get a link to the SerenityOS FAQ",
+            deleteRequest: true,
         },
         {
             name: "build",
             response:
                 "How To Build: https://github.com/SerenityOS/serenity/blob/master/Documentation/BuildInstructions.md",
             help: "get a link to the build docs",
+            deleteRequest: true,
         },
         {
             name: "wsl",
             response:
                 "WSL Specific Notes: https://github.com/SerenityOS/serenity/blob/master/Documentation/NotesOnWSL.md",
             help: "get a link to the wsl specific notes",
+            deleteRequest: true,
         },
         {
             name: "install",
             response:
                 "Installing on real hardware: https://github.com/SerenityOS/serenity/blob/master/Documentation/INSTALL.md",
             help: "get a link to the directions for installing SerenityOS on real hardware",
+            deleteRequest: true,
         },
         {
             name: "bot-src",
             response:
                 "Bot Source: https://github.com/SerenityOS/discord-bot/tree/master/src/commands",
             help: "get a link to the source code for bot commands",
+            deleteRequest: true,
         },
         {
             name: "soytineres",
@@ -39,6 +44,7 @@ export class QuickLinksCommand implements Command {
             response:
                 "https://cdn.discordapp.com/attachments/830522505605283866/834516065517568030/unknown.png ",
             help: "!SOytinereS ot emocleW",
+            deleteRequest: false,
         },
     ];
 
@@ -56,6 +62,7 @@ export class QuickLinksCommand implements Command {
         const link = this.links.find(link => link.name === parsedUserCommand.parsedCommandName);
         assert(link);
         await parsedUserCommand.send(link.response);
+        if (!link.deleteRequest) return;
         try {
             await parsedUserCommand.originalMessage.delete();
         } catch (e) {
